requirements.txt
streamlit==1.28.1
python-docx==0.8.11
pdfplumber==0.10.3
openpyxl>=3.0.0
groq>=0.4.1
